revamp the health check mechanism
Before: we used to check for the number of consumers connected and compare it to an expected number of consumers, if that didn't match the current consumer reports a health problem. This proved to be unreliable as more consumers get connected to the same queue, gradually it starts breaking consumers it shouldn't.

Now: The new way of performing health checks consists of checking the current consumer by its tag. When the consumer connects to the queue, RabbitMQ returns the tag generated for this consumer. Once done, Bowler saves this tag to a file in the app's storage path and uses the Management Plugin to fetch the list of connected consumers to the indicated queue and checks for this consumer's tag among the list.